fix codex

+++ base/sys_event_logger/cache/json_parser.cpp
@@ -76,8 +76,8 @@ void JsonParser::ParseUInt32Vec(const Json::Value& root, std::vector<uint32_t>& 
     if (!root.isArray()) {
         return;
     }
-    for (int i = 0; i < root.size(); ++i) {
-        vec.push_back(ParseUInt32(root[i]));
+    for (size_t i = 0; i < root.size(); ++i) {
+        vec.push_back(ParseUInt32(root[static_cast<int>(i)]));
     }
 }
 
@@ -86,8 +86,8 @@ void JsonParser::ParseStringVec(const Json::Value& root, std::vector<std::string
     if (!root.isArray()) {
         return;
     }
-    for (int i = 0; i < root.size(); ++i) {
-        vec.push_back(ParseString(root[i]));
+    for (size_t i = 0; i < root.size(); ++i) {
+        vec.push_back(ParseString(root[static_cast<int>(i)]));
     }
 }
